requirements.txt
anyio==3.7.1
attrs==25.1.0
Authlib==1.2.0
fastapi==0.87.0
fastapi-versioning==0.10.0
fastapi_versionizer==1.2.0
httpcore==0.16.3
httpx==0.23.3
pytest==8.3.2
pytest-asyncio==0.24.0
pytest-env==1.1.3
pytest-mock ==3.6.1
requests==2.26.0
starlette==0.21.0
urllib3==1.26.20
uvicorn==0.20.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==14.2
motor==3.1.1
pymongo==4.6.3
python-dotenv==1.0.0